# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Clean E-yAy API
  version: 2.0.0
  description: |
    Sözleşme-önce API. Pydantic ve TypeScript tipleri buradan üretilir.
    `make codegen` ile yenilenir. Manuel düzenleme yok.
servers:
  - url: http://127.0.0.1:9000
    description: Local dev
paths:
...
  /api/v1/notifications:
    get:
      operationId: getNotifications
      summary: Bildirim listesi (en yeni önce; observer — karar vermez)
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: unread_only
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"
  /api/v1/notifications/{id}/ack:
    post:
      operationId: ackNotification
      summary: Bildirimi okundu işaretle
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  id: { type: string }
...
  /api/v1/paper-trading/maintenance/archive:
    post:
      operationId: postPaperTradingMaintenanceArchive
      summary: Owner-admin — paper state'i arşivle (read-only kopya; mutasyon yok; paper-safe)
      parameters:
        - name: reason
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK (MaintenanceResult; paper_safe/no_execution; additive)
          content:
            application/json:
              schema:
                type: object
                description: Maintenance sonucu (archive; additive).
  /api/v1/paper-trading/maintenance/repair:
    post:
      operationId: postPaperTradingMaintenanceRepair
      summary: Owner-admin — paper state tutarsızlıklarını onar (dry_run varsayılan true)
      parameters:
        - name: dry_run
          in: query
          required: false
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: OK (MaintenanceResult; dry_run ise yazım yok; non-dry-run önce arşivler)
          content:
            application/json:
              schema:
                type: object
                description: Maintenance sonucu (repair; additive).
  /api/v1/paper-trading/maintenance/reset:
    post:
      operationId: postPaperTradingMaintenanceReset
      summary: Owner-admin — paper hesabı sıfırla (önce arşivler; learning/decision log korunur)
      parameters:
        - name: reason
          in: query
          required: false
          schema: { type: string }
        - name: preserve_learning
          in: query
          required: false
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: OK (MaintenanceResult; arşiv + temiz initial state; broker yok)
          content:
            application/json:
              schema:
                type: object
                description: Maintenance sonucu (reset; additive).
...
components:
  schemas:
...
    Notification:
      type: object
      required: [id, ts, type, priority, title, body_short]
      properties:
        id: { type: string }
        ts: { type: string }
        type:
          type: string
          enum:
            - ticket_created
            - ticket_expiring
            - ticket_expired
            - ticket_blocked
            - recheck_exit_recommend
            - recheck_reduce
            - risk_gate_changed
            - risk_kill_switch
            - catalyst_imminent
            - dqs_dropped
            - position_near_sl
            - position_near_tp
        priority:
          type: string
          enum: [critical, high, medium, low]
        title: { type: string }
        body_short: { type: string }
        body_long: { type: string }
        ticket_id: { type: string, nullable: true }
        position_id: { type: string, nullable: true }
        actions:
          type: array
          items:
            type: object
            additionalProperties: true
        ack: { type: boolean }
    NotificationList:
      type: object
      required: [notifications, unread_count, total]
      properties:
        notifications:
          type: array
          items: { $ref: "#/components/schemas/Notification" }
        unread_count: { type: integer }
        total: { type: integer }
...